Create stub method when lifting multiple constant values
Summary:
When we lift constant values from a method, if there're multiple constant values and more than one call sites to the modified method, the inserted constant loads are likely to pollute the call sites and increase their code size.
To alleviate this problem, this change creates a small stub method that holds the lifted constant values and pass them to the modified callee. We can, therefore, update the original call sites to invoke the new stub instead.
In this way, we leave the inlining decision to SimpleInline to determine the final shape of the call. Small stubs with few call sites will be inlined; Larger stubs with multiple call sites will unlikely to be inlined. This seems to be a better way to balance code size.

Reviewed By: minjang

Differential Revision: D9729775

fbshipit-source-id: 73f9b02007299e74606e1f2c24417672265d4331