Fix some non-determinism in SimpleInline
Summary:
D9390218 causes at least some of the non-determinism we're seeing with inlining. This unordered_map is being iterated on later which determines opcode sizes and what to inline and what not to inline.

Swap in ordered map.

Note that as the repeated runs of statsbot shows, this clearly doesn't fix the problem entirely. Likely one problem of several here?

Reviewed By: thezhangwei

Differential Revision: D9724758

fbshipit-source-id: f19e8f3f356328510a07ba50148e0c18091b7fb0