Fix cloning of parent pointers
Summary:
Context: copying the callee CFG so it can be inlined

Iterating over the blocks in a CFG is done in an arbitrary order, so there's no guarantee that the parent position has already been processed when you come across a parent position pointer. The fix is to go over in two passes, fixing parent pointers on the latter pass.

Reviewed By: emma0303

Differential Revision: D9826306

fbshipit-source-id: fcc0a6e637c105cd63df08ccd7288954f10bbe51