Handle block deletion during inlining
Summary:
Fix a bug where we tried to remove a callsite that was just deleted. See details in the comments of `inline_with_cfg`

Right now, SimpleInline can use either IRCode or the CFG to do the inlining. Which is a bit awkward because the editable CFG takes the code from out of the IRCode object -- they can't both be useful at once.

For now, `inline_with_cfg` builds the cfg and destructs them for every call (quite expensive), but before I land this, I plan to rewrite SimpleInline to use the editable CFG, so this won't be a problem later.

Reviewed By: int3

Differential Revision: D9925233

fbshipit-source-id: 9e2e530f9857f29e968149bbcca81332038c4773