Don't add catch after catch all
Summary:
Before this diff, we were adding throw edges from the callee at any place it could throw, but we forgot to take into account throws that are inside a catch-all, and tried to add catch edges after the catch-all, which doesn't make any sense. There isn't anywhere you could go after a catch-all.

This manifest as a very confusing bug and took me a while to trace this down. Nothing asserted about the malformed data, and it generated dex files. But those dex files were invalid. Dexdump refused to read them. It was complaining about an address that was larger than the code section it referred to. The issue was that the dex file we created put a type index where a code offset should have been (because of the catch-all that wasn't at the end of its list).

I've added asserts that would catch this (and similar) issues in the future.

Reviewed By: int3

Differential Revision: D9925234

fbshipit-source-id: f0728b2343b05577deb5fe206ba91d785d90697e