Properly merge cross referencing ctors
Summary:
It's common that we have multiple ctors defined in a class that reference each other.
```
class A {
public A() {
  A("Foo");
}
private A(String s) { ... }
```
As shown above, the 1st ctor calls the 2nd private one.
When creating the dispatch of the 1st ctor, we inline each ctor entries into the dispatch.
Note that the inlined entries now have calls to the other ctor that we also try to merge.
Therefore, we need to patch those calls properly as well after merging both ctors.

We are not collecting existing call sites to the original ctors correctly. We missed the ones inside the inlined ctor bodies.
By collecting call sites after inlining ctor dispatch entries, we can fix this subtle bug.

Reviewed By: emma0303

Differential Revision: D9997605

fbshipit-source-id: f2b3c84edec834c27b6c569d5b8df3ebd11333e5