add a dex store optimization pass to move classes if they are only referenced by a single store
Summary: When the BUCK code sorts classes into dex stores, it does so on a buck target basis.  Some targets we build are from large third party libraries and are not amenable to splitting up the target so that classes that are exclusive to a dex store are properly included in the dex store.  This pass builds up a reference graph for each class and then moves classes into the proper store if they are exclusively referenced by that store.

Reviewed By: int3

Differential Revision: D9759539

fbshipit-source-id: 93cd2804ba0440b6a4dd6fb5ba8c757968cacee3