Inline small methods regardless of whether they can be deleted
Summary: Even if we can't delete a method, it's still worth inlining if it has `<= 2` instructions. One of those instructions is usually a return (only other case is if the method ends in a throw), and since inlining elides that return, we are basically just replacing an invoke instruction with another one that is almost always the same size or less. We don't lose on code size, and we benefit from making intraprocedural optimizations more effective -- e.g. inlining methods that return a constant values means that intraprocedural constant propagation can elide branches on their return value. We may also reduce cross-dex references slightly.

Reviewed By: emma0303

Differential Revision: D10010145

fbshipit-source-id: 9553dbe3d511e83abee368947f3229d996d3eb65