Don't mutate PG keep rules when processing keepclasseswithmembers
Summary:
Background: I wanted to make sure `process_proguard_rules` was idempotent because I wanted to have the option of re-running it and asking it to spit out more detailed diagnostics (in particular, to return for each keep rule the set of classes matched). As such, I tried to make all the inputs to `process_proguard_rules` const, whereupon I noticed that `MemberSpecification::mark_conditionally` was being used in a non-const way.

`mark_conditionally` was being used to handle the `keepclasseswithmembers` PG rule. On closer look, I saw that a *lot* of the logic around the handling of this rule was unnecessary. To handle something like

```
-keepclasseswithmembers class * {
  void foo*();
  void bar*();
}
```

we just need to find all classes with at least one member starting with `foo` and one starting with `bar`.

This diff simplifies that code as well as making `const` all the inputs to `process_proguard_rules`.

Reviewed By: justinjhendrick

Differential Revision: D10056352

fbshipit-source-id: 2d676b3a63faeeaf87eeb6459e237a997e6d1b4e