Faster method sorting by skipping copies, and use fully deobfuscated name
Summary:
The sort step in D9764772 is quite slow- I found it added in the order of minutes (as much as 10-20) onto a release build. This is because the method comparison map is getting copied around. By switching to pointers, we avoid that cost (copying pointers is much cheaper than copying the full map). This reduces the cost of sorting the primary dex from ~43,000ms to ~383ms.

Also switched from 'get_deobfuscated_name' to 'get_fully_deobfuscated_name', which handles cases like methods that were created by redex. This appears to have sorted a decent number (2000+) of additional methods to the top of the first secondary dex.

Reviewed By: stashuk

Differential Revision: D10133808

fbshipit-source-id: 4c68d7f9ad94985258fbed9f962c24bfaf07e27c