Exclude mergeables referencing android sdk types
Summary:
The motivation is we were most likely merging classes referencing android sdk types that do not exist on certain android versions. This complication causes the shape not to verify at runtime. Which in turn triggered additional class loading during cold start. Therefore, we want to avoid merging classes that have such references.

Here we add an optional check for all mergeables in a model to exclude the ones that have reference to android sdk.
The feature is driven by a small JSON cluster in the config. We can exempt android sdk types from the check as long as we know it's safe to do so.

The Redex runtime impact is minimal since it only runs for the enabled models.

Reviewed By: emma0303

Differential Revision: D10227498

fbshipit-source-id: 96740a9dcc595cd0c120cf9503c7991ed9c89c2d