nextPowerOfTwo => roundUpToPowerOfTwo
Summary:
There was a misinterpretation of the oat type table spec in oatmeal; the table should be the nearest >= power of 2 for N types, but when N was exact power of 2, we were sizing the table at the next higher power of two. Fix that, rename the function to be more clear about its spec, and add unit tests.

Note, there was one use of this function outside the type table, and I think it's fine to use the new semantics in that case, but not entirely sure 100% how to verify this.

Reviewed By: emma0303

Differential Revision: D10385976

fbshipit-source-id: d24a9f615d75b2f005ee72b8ab56347bf533bac0