Check if the method logged is in the current dex when write method id map
Summary:
in write_classes_to_dex(...) in DexOutput, we use gather_methods to gather methods that classes in current dex referencing to. So method was recorded in dex it was defined at, and also dexes it was referenced at.

To generate correct encoded method-id for debug line map (and also reduce unnecessary method idx and dex signature recorded in method id map), check if the method's class is in the current dex before recording method idx and dex signature for the method.

Reviewed By: int3

Differential Revision: D10389908

fbshipit-source-id: 5f995ba9638d5ed39fe5601474b10352b30f2b0e