Use Editable CFG Adapter to run SimpleInline quickly
Summary:
I created a config flag in SImpleInline called `use_cfg_inliner` for the transition from IRCode to editable CFG based computation. The original implementation of `use_cfg_inliner` was very slow. It built and destroyed both CFGs every time a method was inlined, without parallelism.

Before this diff, If a method was inlined multiple times, we rebuilt its CFG multiple times :(. In this diff, we build all the CFGs in parallel at the beginning and destruct them all in parallel at the end. Which is much faster (~2x). Though, this is still slower than the IRCode implementation :/.

This diff uses the Editable CFG Adapter (created in previous diff) to share code with the IRCode implementation. This way, we have the `use_cfg_inliner` flag to toggle in case there are bugs.

Reviewed By: int3

Differential Revision: D10452816

fbshipit-source-id: b2bdc31f0388b9fbbc2abc4063523a3ea010fe61