Reverting defaults for string_sort_mode
Summary:
The string-mode sort by class reference order causes
performance/IO problems with the internal ART class loader
where we're not using the locator strings.  It also screws
with compression algorithms for compressing the dex.
This change showed a 750kB win on APK size on
FB4A, so moving it more broadly.

Reviewed By: minjang

Differential Revision: D10852093

fbshipit-source-id: c8c107a7b546f0c095fafc99b7732d1964c29118