Fix missing line numbers in inliner
Summary:
Fix a bug where Logview reported the wrong inlined method when two methods were inlined into a single caller. This bug happened because one of the callees didn't have any line numbers in the final apk, it was placed after the other inlined method and inherited the last line number from the other method.

We first inline the callee code until its return statement, then we put a position corresponding to the callsite at the end of that chunk of callee code. The rest of the callee code (after its return statement) is placed at the end of the caller method. This was a problem because the chunk of code after the return may not have any positions in it. It will inherit the position of whatever code comes before it.

To fix this problem, we copy the last position before the return in the callee to the first position after the return.

See `SimpleInlineTest.cpp` for an example

Reviewed By: minjang

Differential Revision: D10500339

fbshipit-source-id: ed0479f13130ea0a5fcf072e532a182d97dfd5ce