Avoid updating the super class of an external type
Summary:
When updating the super class of an interface rooted mergeable type, the implementor of the interface could be an external type.
There's no need to touch external types, since they don't exist in the apk.
The diff adds an additional check for external types when updating the super class for an interface root's implementors.

Reviewed By: emma0303

Differential Revision: D12853362

fbshipit-source-id: 3d11a0ccc13d2342c279409a73d789f290fd703d