Speed up construction by avoiding src_is_wide
Summary:
Speed up `build_cfg()` by about 10%.

We do this by speeding up `compute_registers_size()`. `compute_registers_size()` called an expensive function `src_is_wide()` because it was inspecting the prototype on `invoke`s. I realized it's not actually necessary to check the source registers in well-formed input code. We should never see a register used as a source that isn't also used as a destination register.

However, this adds a new constraint to test code, we can't use an undefined register in tests anymore :/

Reviewed By: int3

Differential Revision: D12812358

fbshipit-source-id: eeb5c1fd6f19c547c990576910249c6ee91d0d97