Add support for parent in .pos directive
Summary:
At a high level I support parent by encoding the index of the parent position that is referred to.

When serializing I recursively serialize parents before the current pos is encoded so that (a) parents always exist (they will only be force encoded this way if somehow we were handed IRCode that references a parent outside itself) (b) parents always appear before their children.

Deserializing is easier since we just error if the parent is out of range

Reviewed By: int3

Differential Revision: D12968969

fbshipit-source-id: 924cfaea24ab06495180781a53636bb9d6c0b262