Add root type name tag in the name of the generated Shape class
Summary:
When combining multiple roots in the same TE Model, all the generated shape classes share the same name scheme.
They are all based on the prefix name of the model config. This naming is confusing sometimes, since we don't know from which root the shape is generated from.

To address this confusion, we add another short name tag on the shape class based on the class name of the root type.
We also shorten the prefixes. In this way, the name of the generated shapes is more concise.

Reviewed By: emma0303

Differential Revision: D12912194

fbshipit-source-id: 8e8c48ebd0bce3e57c66d2ef9e14836ca942ebe6