Introduce Enhanced TypeEnvironment
Summary:
The current TypeEnvironment only tracks the type of a register using a simple type lattice. This is not enough to distinguish between reference types. We need this more precise type information to become feature parity with the Android verifier.

This is the very first step. We add EnhancedTypeEnvironment which augments the previous TypeEnvironment with precise type information only when the basic type is a reference type.

This diff only scaffolds this ReducedProductAbstractDomain EnhancedTypeEnvironment. Right now it simply delegate all operations to the first PatriciaTree component mapping to the simple type lattice. The actual precise type operation logic is coming in the future diffs.

Reviewed By: int3, emma0303

Differential Revision: D12935763

fbshipit-source-id: 806191229c467957eb7d485215b071c34a608b55