LayoutReachabilityPass: Recompute XML-to-code references after XML optimization passes
Summary:
We have a fair bit of dead code kept around by dead XML references. While we have passes that remove the dead XML refs, we weren't removing the corresponding bits of dead code. This diff does that.

wsanville had an earlier attempt in D9139758, the code in this diff is actually mostly lifted from that.

Reviewed By: wsanville

Differential Revision: D13037192

fbshipit-source-id: e708f7b6ec031eb2428cca326a619fcf5d51a11a