Locate classes by implied index
Summary:
Previously, the renaming of classes performed by Redex was independent of the way the 2 class loaders locate classes at runtime.

For each reference to a non-framework type (including array references), a separate locator string was sneaked into the string table just before the actual type descriptor string. This locator string contains a base-94 encoding of a (store number, dex number, class number) triple. Given that Fb4a has lots of type references, all those locator strings add up.

Meanwhile, the class renamer in Redex assigned names of the form LX/nnnnnn; to all classes that can be freely renamed, where nnnnnn is a base-62 encoding of a increasing sequence of number.

This diffs gets rid of the need of encoding separate locator strings for renamed classes.

To this end...
- We make sure that the LX/nnnnnn; scheme doesn't just represent any increasing sequence of numbers, but that in fact nnnnnn represents the exact class's class table index plus some dex-specific offset. We call such an nnnnnn number the global class index. Each dex defines class with global class indices in a unique range (different from the ranges of other dexes).
- We introduce a new option emit_name_based_locator_strings
- If the option is set, we store a description of that range in each dex file (similar to how locators were stored in the past, we sneak this information in just before the empty string, which is always present and easy to locate at index 0).
- If the option is set, we also modify the metadata.txt file to include a line ".emit_name_based_locator_strings"
- We modify the class loaders...
  - to detect if the ".emit_name_based_locator_strings" metadata line is present, and if so,
  - to decode the dex range information and keep it around, and
  - whenever trying to resolve a type name, first try to see if the type name represents a global class index, and then find the defining dex and the class index using the dex's range information
  - since the class loaders in dextricks are on a different shipping cycle, the new behavior is guarded and only active if the dex manifest contains a .name_based_locators entry

We still fall back to the old locator scheme for non-renamed classes.

In the class loader, the new scheme brings with it a single locator-string read for each dex at start up time (to learn about dex index ranges), while also getting rid of the subsequent need to read locator strings from the string table for every single class load event for renamed classes.

Because of how the Redex release process works, and how buck is set up, the unpacker.py has to be made aware of the new option across all Redex release branches.

The legacy RenameClassesPass has not been updated, and shouldn't be used together with emit_name_based_locator_strings; Redex would now abort with an error in that case. This combination seems only to be configured in fb4a_instr.config; I am not sure if/where that configuration is actually used.

Reviewed By: tophyr

Differential Revision: D12897490

fbshipit-source-id: 35e5233334249d20e88db903b5a0a63378452816