Strip debug locations if the following instruction doesn't throw
Summary:
See title. I do this in a second pass to keep the first pass in StripDebugInfo less complicated

At a high level the algo is "iterate backwards, find non-throwing instructions, prune any positions that appear before non-throwing instructions, with the expection of positions that are parents of other positions". We assume that parent positions appear before their children (which appears to be true by the code I've read from the inliner).

Reviewed By: justinjhendrick

Differential Revision: D12947492

fbshipit-source-id: 0aeaf8731b842d9c9ff6910f2dd660be6738d9eb