Remove redundant check casts
Summary:
As title. This analysis provides a proper way to check for redundant checks (and will replace the pattern matching one).

NOTE: This will also make the proper fix for the SingleImpl bug easier to fix (as soon as I get to it).

and

NOTE: This could be replaced in the end with thezhangwei's analysis for types, but for now that is still WIP and more complicated that needed here.

Reviewed By: arnaudvenet

Differential Revision: D13080086

fbshipit-source-id: 43acb8f3549a5be5cb0a0b94ff2399ff04a7aa8c