sign extend dex encoded floats
Summary:
I found this via a bug where a float was being encoded as a signed 64 bit integer. It had all 0s in the upper 32 bits and a 1 in the 32nd place, So when we right shifted it, the upper bits got filled in with 1s and signed_int_fits<32>(literal) failed.

follow-up to D12851712

Reviewed By: int3, minjang

Differential Revision: D13074425

fbshipit-source-id: 1e332503ee170a9212ce5bf079d48336a1d85c0f