Fix line numbers in CFG
Summary:
DexPositions apply to every instruction in the linear stream until the next DexPosition. We had a bug in the CFG where we would split the code into blocks and re-order the blocks without considering which DexPosition each instruction should belong to.

This diff fixes the problem by adding a DexPosition to the beginning of each block, if it doesn't have one already. At linearize time, we clean up duplicated DexPositions.

This diff introduces a ~50k regression, but the next diff in the stack fixes it. (The whole stack is a small win). The reason is that this diff creates more blocks that only have positions in them. In the next diff, `simplify()` is improved to delete those blocks.

Reviewed By: danzimm

Differential Revision: D12889073

fbshipit-source-id: 4b5769ac1e7b2c0428314c56ec77e10315fb044c