no empty blocks
Summary:
Make the CFG delete blocks that only have positions in them. Previously, it was only deleting completely empty blocks.

This is required because there's a bug in dex2oat (version 6.0.0_r1) that generates bogus machine code when there is an empty block (a block with only a goto in it). To avoid this bug, we use the CFG to remove empty blocks.

Reviewed By: tophyr, danzimm

Differential Revision: D13223065

fbshipit-source-id: d867487bfd4001e1d7b20956769509be2717fb49