Proper true virtuals detection when updating colliding method signatures
Summary:
The way we detect true virtual when handling colliding method signature update was wrong.
We constructed the virtual scope after we partially updated the signatures. At that point, we already broke the virtual scope of the program. We were in a transient state. A virtual scope computed based on it is invalid.

The proposed fix is to introduce one more step that only update non-colliding ctor and vmethods. We do so after the virtual scope computation. If no colliding candidates were found we simply update and return.

Reviewed By: emma0303

Differential Revision: D13385656

fbshipit-source-id: 94611ff330a012fa73c965afacc1855ec43b3855