Ordering classes to reduce number of cross dex method+field refs, strings and types
Summary:
The classes in the first few dexes are special (because of primary dex,
mixed mode, and interdex / beta maps). However, the vast majority of the code
of large apps lives in the remaining dexes. It used to be the case that Redex
made no significant changes to the ordering of the classes (besides type
erasure) in those remaining dexes, leaving the order as computed by dx.

This diffs changes that: We now maintain a set of the remaining dex classes
with associated priorities based on the *ref needs of the classes and the
*refs already added to the current dex, and always pick the class with the
highest priority to emit next.

The priority of each class is roughly determined as follows.
- The primary priority is given by the ratio of already applied *refs to
  unapplied *refs. This ratio is slightly tweaked in favor of
  singleton *refs. ("Applied" refs are those which have already been added
  to the current dex. "Singleton" refs are those for which there is only
  one, namely this, class left.)
- If there is a tie, use the original ordering as a tie breaker

(All this isn't entirely accurate, as it doesn't account for the dynamic
behavior of plugins yet. The prioritization scheme is a heuristic anyway
with lots of room for improvement.)

Reviewed By: int3

Differential Revision: D13251712

fbshipit-source-id: 587be22352a95ec514065341d8f3c18ed7385b0a