Update non colliding ctor in the first step to detect subsequent ctor collisions
Summary:
In the first step of updating type reference in method protos, we should update ctors right away if no existing collision is found.
This is required. Since the update is needed to detect a subsequent collision to the same target proto.
When we made the previous change we messed up this logic.

Reviewed By: minjang

Differential Revision: D13499722

fbshipit-source-id: 18b506364d3011cc350b732e91cee0117e47d9b7