When splitting merging group using max_count avoid creating a last group of size one
Summary:
We don't want to create a merger for a single input mergeable class. That's not very useful.
However, when splitting a merging group using `max_count` there's a corner case where we could leave with a last subgroup of size one.
We should avoid this by simply combine the last two subgroups.

Reviewed By: emma0303

Differential Revision: D13503755

fbshipit-source-id: d8cd727a8ff1aa5dd25eedecb6ee2597829e4463