Reland "[redex][check breadcrumbs] Don't fail on references between depending modules"
Summary:
I've had to revert the previous change in D13528296.
Seems like some redexunstable builds were failing, because the type that was passed to `illegal_ref` wasn't in any store, and then `type_store_idx` pointed out of the array, leading to a crash. Here I've added these lines, to fallback to previous behavior if this case happens.
  if ((store_idx >= m_xstores.size()) || (type_store_idx >= m_xstores.size())) {
    return false;
  }
I'd like to reland it, as I think this diff does fixes some bugs around the way we're deciding if a ref is illegal, specifically it fixes a few cases where `SingleImpl` was inlining illegally.
Original commit changeset: c35b815b6dcf

Reviewed By: justinjhendrick

Differential Revision: D13545039

fbshipit-source-id: 02deee74daace75d0312ff4caab1693830566e08