set type alias in DexClass::set_deobfuscated_name
Summary:
I want to be able to look up a type by its deobfuscated name or its real name. This diff makes `set_deobfuscated_name` create an alias to the DexType.

The previously existing function `assign_name_alias` wasn't actually creating an alias as I expected, it was resetting the name rather than creating an additional name that maps to the same type. So instead, I created a `set_name` to do the old behavior, and a new function called `alias` that assigns an additional name to a type (for deobfuscated names).

Reviewed By: int3

Differential Revision: D13582888

fbshipit-source-id: c78da10cc48473b8c29ce2353b88b003f4758435