dont inline across TargetApi boundaries
Summary:
In D13469164, I only did `RequiresApi`, but `TargetApi` has a very similar purpose and the inliner shouldn't inline across that either.

This diff also reads the min-sdk from the manifest file. All code that doesn't have an annotated sdk level is treated as having the min-sdk level.

In an attempt to minimize the loss of this diff, I included un-annotated inner classes into the api level of their outer class. This isn't strictly correct but it seems like most developers just forgot (or were too lazy) to put the annotation on all the inner classes. The way I did it is kinda sketchy though... I just removed everything after the last `$` then looked for a class with that name :/

Reviewed By: int3

Differential Revision: D13553911

fbshipit-source-id: da2eb8b1fb719873cee0d7dc19a9bc06c8d78178