Be smarter about which manifest references are reachability roots
Summary:
We mark an <activity>'s referenced class as reachable only if it is exported
or has intent filters. Exported Activities may be called from other apps, so
we must treat them as entry points. Activities with intent filters can be
called via implicit intents, and it is difficult to statically determine
which Activity an implicit intent will resolve to, so we treat all potential
recipient Activities as always reachable. For more details, see:

  https://developer.android.com/guide/topics/manifest/activity-element
  https://developer.android.com/guide/components/intents-filters

Note: RMU may delete some of the Activities that we haven't marked as entry
points. However, it currently doesn't know how to rewrite the manifest to
remove the corresponding <activity> tags. This seems benign: the Android
runtime appears to be OK with these dangling references.

This diff has quite a bit of refactoring because I did not want to couple the
XML walking code with complicated reachability logic. Instead, I got the XML
walker to gather the necessary pieces of information in a convenient
`ManifestClassInfo` struct which the reachability analysis then uses.

Reviewed By: wsanville

Differential Revision: D13597853

fbshipit-source-id: fb9feff07251ab5fb1c7e94d71cbb1a0087468e0