In cross-dex minimization strategy, favor classes with infrequent references
Summary:
This diff generalizes the "singleton" concept of the cross-dex minimization strategy.
Before, the selection process would already favor classes which have references that other no other remaining class has. This is a good idea, because no other dex will then also have to include that reference.

The generalization of the concept now guides the selection process towards getting into a state where more such singleton references exist and can be taken advantage of.

This is done by not only tracking singleton references (those with only one class having that reference), but also references where two classes have that reference, and also references where three classes have that reference, and so on.

This information is the used as part of the priority function that guides class selection: Tend to give somewhat higher priority to classes that have such infrequent references, boosting the priority more for more infrequent references.

Also, this diff introduces the notion of a "weight" for references (instead of effectively weight 1).

Reviewed By: emma0303

Differential Revision: D13555810

fbshipit-source-id: 4181139efbdbea4cec3f6ed10317cc7617dbd026