Partially accounting for classes merged by type erasure
Summary:
Type erasure works via the inter-dex plugin in a special way: Effectively, when a class is emitted that references mergable classes, then those mergable classes will end up in the same dex.

This diff ensures that this dex affinity of the mergable classes is reflected in the prioritization strategy, influencing further class selection knowing that those mergable classes will already be present in the dex.

This is still "reactive", in that that prioritization strategy isn't a-priori away of that certain classes will drag other mergable classes in a dex, but at least the references of type-erased classes are properly accounted for.

Further improvements may include...
- Making the prioritization strategy aware of which classes will bring along what other classes
- The actual class merging should get reflected in the *refs sets that the cross-dex minimization strategy maintains.
- I only handle the type erasure plug-in here; something similar should probably happen for the SwitchInline plugin.

Reviewed By: emma0303

Differential Revision: D13559333

fbshipit-source-id: c2768da66965dc9143afe61d6bea816e48e19914