Emit metadata for the symbolicator to use when symbolicating traces
Summary:
There's a comment in code explaining why iodi metadata is necessary, but I'll add another summary here:

Currently when android prints a stack trace line it does so by printing `"{pretty_class}.{method}"` where `pretty_class` is the class type name with `L,;` removed and all `.` replaced with `/` and `method` is the string pointed to by the method_id_item for the given method. In particular this means that multiple methods may show up with the same stack trace line (for example, if a method was overloaded or if templating was used). Previously these methods were de-collided using the proguard mapping; the mapping contains line number ranges for some methods so that when symbolicating the symbolicator looks at the current line number to determine which method is being represented. With IODI we no longer can leverage this mapping because the line numbers emitted are now instruction offsets.

In order to solve this problem we emit a metadata file that will aid in symbolication. Symbolication will work as follows:

- If there is no collision with another method name then we have a 1-1 mapping between a method's pretty string and its method_id.
- If there is a collision, but it's to a method that can be renamed then we rename it such that there is no collision and we fall back in the first case.
- For direct/static methods (which are the majority of non-renamable collisions) we additionally emit a list of callers and the pc they call said callee at. During symbolication, then, we look at the previous frame's method id and pc to indicate which initializer is being called.
- For any other methods that don't fall into the above cases we emit usual debug info.

In theory in the future we can take the third bullet point further for any other colliding methods pretty names that don't override an external interface/virtual method (otherwise in theory the previous frame might come from an external library that we don't emit metadata about), but for now that is a non-goal.

Below is pseudo code for the proposed symbolication of an IODI trace:
```
trace: trace_t
result: list<(method_id, pc)>
for line in trace.reverse():
  if line.pretty_method in iodi_metadata.non_colliding:
    result.append(iodi_metadata.non_colliding[line.pretty_method], line.line_number)
  elif line.pretty_method in iodi_metadata.colliding:
    pc = line.line_number
    previous_method_id, previous_pc = result[-1]
    if previous_method_id == 0: # failed to symbolicate last line, fuzz it
      # Look at all possible callers and try to match the pc...
    if previous_method_id != 0:
      collision_meta = iodi_metadata.colliding[line.pretty_method]
      result.append(collision_meta[previous_method_id][previous_pc], pc)
    else:
      # We failed to find the previous method_id, so make it 0 but append
      # pc still so we can try to fuzz
      result.append(0, pc)
  else:
    result.append(symbolicate_normal(line))
return map(lambda p: symbolicate_lean(p), result.reverse())
```

Where `symbolicate_lean` uses the already built infrastructure to map from `(method_id, offset) -> (file, line)` and `symbolicate_normal` uses already built infra to symbolicate a line as if it has normal debug info.

First let's break down changes per file:

- `IODIMetadata.{h,cpp}`: These files compute and encode the metadata described above
- `iodi.py`: This decodes the metadata into json (if invoked from the cli) or a python object (if loaded as a module)
- `redex-all/main.cpp`: This needed to change to invoke IODIMetadata construction & writing if the `iodi-metadata` key was passed
- `DexOutput.{h,cpp}`: This changed so that we fell back to emitting normal metadata if we cannot safely use IODI for a given method (if it collides with another method and is a virtual/interface method)
- `Trace.h`: I wanted to add some IODI trace statements, so I added a tag for it here
- `redex.py`: Now we need to copy iodi-metadata if it was created

NOTE: For now we only use this for callees that are direct/static. This is due to the fact that we can guarantee that no external frame will be before ours for direct/static methods (but we cannot guarentee that for any random invocation of an interface/virtual method).

Reviewed By: justinjhendrick

Differential Revision: D13558789

fbshipit-source-id: f344cb15f37ee40f8efad1dc67259d73e3a68781