Add output file to RemoveUnreachablePass with removed classes, methods, fields
Summary:
Produce an artifact that stores the symbols of all removed code objects.

This is useful for dead code removal - if a codebase is used to build multiple apps, and a code object is unreachable in each one, then that code object is safe for deletion. RemoveUnreachablePass runs multiple times, but this output is only likely to be human-meaningful on the first pass.

Reviewed By: int3

Differential Revision: D13630174

fbshipit-source-id: 01500c52a42a020b12b271b9ad0f2f0708425893