Merging cross-dex-ref-minimization feature into redex-stable
Summary:
This is to facilitate A/B testing.

This pulls in D13411751 (an InterDex pass bug fix), and the cross-dex ref minimization feature (D13251712, D13555810, D13557704, D13559333, D13557687, D13600989). All diffs applied cleanly with the following commands.

  cd ~/fbsource
  for diff_id in D13411751 D13251712 D13555810 D13557704 D13559333 D13557687 D13600989; do hg diff --root fbandroid/native/redex -c $diff_id | patch -p1 -d fbandroid/native/redex-stable; done

The only manual change on top of the above mechanical import of redex (unstable) changes is that the config changes in
  fbandroid/native/redex-stable/facebook/config/fb4a.config
  fbandroid/native/redex-stable/facebook/config/instagram.config
have been commented out, so that the new feature isn't enabled by default.

Reviewed By: emma0303

Differential Revision: D13655786

fbshipit-source-id: 82e3066bda89bbd3bdc6ee413ba5084fdf4df37d