Add unittests for SimpleReflectionAnalysis
Summary:
- Setup unittest for `SimpleReflectionAnalysis`
- Add `has_found_reflection` to scan the post analysis environment to simplify reflection check
- Add enum `ClassObjectSource` as an additional field in `AbstractObject` to distinguish between an input class object and an output class object
- More tests to come

Note that the join operation of `ClassObjectSource`s is not correct. It seems like the solution should be turning it into a Lattice  abstracted by a `FiniteAbstractDomain` and combined it with the existing `AbstractObjectDomain` using the `ReducedProductAbstractDomain`. To not mess with the existing diff stack. I will make the change landing the current changes.

Reviewed By: minjang

Differential Revision: D13829264

fbshipit-source-id: 2199bc19897680e89cdb9126e83e8dd504abedb9