Handle if-else tree in SwitchMap
Summary: SwitchMethodPartitioning was originally designed for methods that had a single switch statement, but this diff extends it to support methods that use an if-else tree to choose a case block (instead of a switch). These methods may have been switch-only in source code, but have been compiled into if-else trees (usually by d8).

Reviewed By: int3

Differential Revision: D13872279

fbshipit-source-id: 8d403827d34db2610fb53811fdf6d8189e71ad25